test für conv layer läuft durch

+++ tensor.py
@@ -3,6 +3,6 @@
 # Tensor klasse zum Abspeichern von Daten
 
 class Tensor:
-    def __init__(self, elements, dtype=np.float64):
+    def __init__(self, elements):
         self.elements = elements
         self.deltas = np.zeros_like(self.elements)
